pyproject.toml
[project]
name = "donation-portal"
version = "2.0.0"
description = "Effective Altruism Australia's donation portal"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amqp==5.3.1",
    "arrow==1.3.0",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "billiard==4.2.1",
    "celery==5.4.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "cryptography==44.0.2",
    "Django==5.1.10",
    "django-cors-headers==4.7.0",
    "django-enumfields==2.1.1",
    "django-ipware==7.0.1",
    "django-paypal==2.1",
    "django-pinpayments==1.1.0",
    "django-pymemcache==1.0.0",
    "django-reversion==5.1.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "kombu==5.5.2",
    "mailchimp3==3.0.21",
    "mock==5.2.0",
    "factory-boy==3.2.1",
    "oauthlib==3.2.2",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "pillow==11.1.0",
    "pluggy==1.5.0",
    "prompt_toolkit==3.0.50",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "PyJWT==2.10.1",
    "pymemcache==4.0.0",
    "PySocks==1.7.1",
    "pytest==8.3.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-ipware==3.0.0",
    "python-postmark==0.6.1",
    "pytz==2025.2",
    "pyxero==0.9.4",
    "q==2.7",
    "raven==6.10.0",
    "redis==5.2.1",
    "requests==2.32.4",
    "requests-oauthlib==1.3.1",
    "rratelimit==0.0.5",
    "selenium==4.30.0",
    "setuptools==78.1.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlparse==0.5.3",
    "stripe==12.0.0",
    "trio==0.29.0",
    "trio-websocket==0.12.2",
    "types-python-dateutil==2.9.0.20241206",
    "typing_extensions==4.13.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
    "XlsxWriter==3.2.2",
    "watchdog>=6.0.0",
    "django-debug-toolbar>=5.2.0",
    "pytest-playwright>=0.7.0",
    "playwright>=1.52.0",
    "weasyprint>=65.1",
    "sentry-sdk[django]>=2.34.0",
]

[tool.pytest.ini_options]
testpaths = ["frontend/tests_python"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--tb=short -v"
asyncio_mode = "auto"
markers = [
    "playwright: mark test as playwright test"
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
